Move adjudication test cases away from filter test cases
